Tools: Moved db2 metadata structures to a location shared by all tools
