+++ src/tools/vmap4_extractor/gameobject_extract.cpp
@@ -18,8 +18,8 @@
 
 #include "adtfile.h"
 #include "DB2CascFileSource.h"
-#include "DB2Meta.h"
 #include "Errors.h"
+#include "ExtractorDB2LoadInfo.h"
 #include "model.h"
 #include "StringFormat.h"
 #include "vmapexport.h"
@@ -56,32 +56,6 @@ bool ExtractSingleModel(std::string& fname)
 }
 
 extern CASC::StorageHandle CascStorage;
-
-struct GameobjectDisplayInfoLoadInfo
-{
-    static DB2FileLoadInfo const* Instance()
-    {
-        static DB2FieldMeta const fields[] =
-        {
-            { false, FT_INT, "ID" },
-            { false, FT_INT, "FileDataID" },
-            { false, FT_FLOAT, "GeoBoxMinX" },
-            { false, FT_FLOAT, "GeoBoxMinY" },
-            { false, FT_FLOAT, "GeoBoxMinZ" },
-            { false, FT_FLOAT, "GeoBoxMaxX" },
-            { false, FT_FLOAT, "GeoBoxMaxY" },
-            { false, FT_FLOAT, "GeoBoxMaxZ" },
-            { false, FT_FLOAT, "OverrideLootEffectScale" },
-            { false, FT_FLOAT, "OverrideNameScale" },
-            { false, FT_SHORT, "ObjectEffectPackageID" },
-        };
-        static char const* types = "ifffh";
-        static uint8 const arraySizes[5] = { 1, 6, 1, 1, 1 };
-        static DB2Meta const meta(-1, 5, 0x9F2098D1, types, arraySizes, -1);
-        static DB2FileLoadInfo const loadInfo(&fields[0], std::extent<decltype(fields)>::value, &meta);
-        return &loadInfo;
-    }
-};
 
 enum ModelTypes : uint32
 {
